Append radio group name to radio ids
If a wizard had multiple radios with the same id (for example the Policy
Wizard with multiple templates), or potentially if a page had multiple
wizards, then clicking on the label text would change the setting of the
first radio with that id on the page, instead of the input it was likely
supposed to be associated with.

This makes the input's `id` (and the label's `for`) attribute more
unique, and updates the relevant test to ensure the label is clicked,
instead of the input directly.

Refs:
 - https://github.com/stolostron/backlog/issues/25009

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>